added movies credits api, repository functions, cast list in viewmodel, simple cast row in detail view (need to implement)

+++ app/src/main/java/com/example/mymovies/data/remote/response/movie_images/LogoDto.kt
@@ -5,7 +5,7 @@ data class LogoDto(
     val file_path: String?,
     val height: Int?,
     val iso_639_1: String?,
-    val vote_average: Int?,
+    val vote_average: Double?,
     val vote_count: Int?,
     val width: Int?
 )
